Add some echo to Sandstorm scripts.

+++ .sandstorm/build.sh
@@ -3,11 +3,17 @@
 # This script only runs once, when the app connects to sandstorm.
 set -euo pipefail
 
-
+echo "In build.sh"
 
 cd /opt/app
 
-cp .env.sandstorm .env
+if [ -f /opt/app/.env.sandstorm ] ; then
+	echo "No sandstorm env file, copy."
+	cp .env.sandstorm .env
+else
+	echo "Sandstorm env file exists already."
+fi
+
 
 if [ -f /opt/app/composer.json ] ; then
     if [ ! -f composer.phar ] ; then
